Move `example_project_configuration.py` to `test_helpers`.
