+++ metricflow-semantics/tests_metricflow_semantics/query/test_query_parser.py
@@ -30,11 +30,10 @@ from metricflow_semantics.specs.spec_classes import (
     OrderBySpec,
     TimeDimensionSpec,
 )
-from metricflow_semantics.test_helpers.metric_time_dimension import MTD
-
-from tests_metricflow_semantics.model.example_project_configuration import (
+from metricflow_semantics.test_helpers.example_project_configuration import (
     EXAMPLE_PROJECT_CONFIGURATION_YAML_CONFIG_FILE,
 )
+from metricflow_semantics.test_helpers.metric_time_dimension import MTD
 
 logger = logging.getLogger(__name__)
